Fix test for properly provisioned device.
testFallback could fail incorrectly on devices that are provisioned with
production certificates. This is due to the fact that the remote
provisioning service uses the same root key certificate currently as the
batch provisioned factory keys. This will change in the future, but for
now, match against an intermediate certificate instead. Intermediates
are guaranteed to be different between remotely provisioned and factory
provisioned keys in all cases.

Fixes: 229423779
Test: atest RemoteProvisionerUnitTests
Change-Id: I359a42f64759b36d34d96d078d9a91d8051c32df