Remove metered vs unmetered network checks
Stop checking whether the current network connection is metered or
unmetered when deciding whether to reprovision the attestation key pool.
Currently, metered connections will cause the PeriodicProvisioner to
choose a much shorter expiration period as opposed to a longer lookahead
for unmetered connections. This was done to try to prevent using metered
data when possible.

However, in practice this is unlikely to make an effective difference,
and is likely causing impacts to users that only use network data as
opposed to unmetered networks. At the most, this would be saving
kilobytes of data.

Instead, this CL institutes a cap of data that can be used in a given 24
hour period in the event that errors are occurring during the
provisioning flow with the server. This is done to make sure that data
budgets are respected even in worst case error scenarios.

This patch also happens to fix a bug where GenerateRkpKeyService was not
properly looking into the future to check expiring certificates.

Fixes: 230603861
Test: atest RemoteProvisionerUnitTests
Change-Id: I1bac418587c4161b5e617082fd2f6b622f9a3c9c