Enable crystall benchmarks for RemoteProvisioner tests
Adds a basic crystalball performance test to evaluate the round trip
performance from the device to request and obtain certificates from the remote key provisioning server. Test does include the setup time needed on the device to create the Certificate Request. The test is run normally using atest, and can be setup to run on multiple platforms with forrest.

Test: atest RemoteProvisionerPerfTests
Change-Id: I64e2ca3bb5332078ec08b4ab7e84a4a23b334617